Copy and free buffers received from dr{mp3,wav,flac}.
Assigning the buffer from the dr libraries to the buffer variable, and then using delete on it, is a memory bug that can lead to leaks and crashes.

By using memcpy on these buffers, then calling the correct dr{mp3,wav,flac}_free() method on them, we can take proper ownership and ensure that delete is the correct way to dispose of the buffer, whatever kind of file it was loaded from.

I also changed the name of one of the variables in the mp3 code path to match the parameter name from drmp3.